requirements.txt
openai
requests
beautifulsoup4
pyyaml
ddgs
python-dotenv